server/requirements.txt
Django>=6.0,<7.0
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
drf-nested-routers
pandas
Pillow
openpyxl


